+++ src/main/java/com/unifacef/app_nutricionista/service/AvaliacaoCorporalService.java
@@ -2,6 +2,7 @@ package com.unifacef.app_nutricionista.service;
 
 import com.unifacef.app_nutricionista.model.AvaliacaoCorporal;
 import com.unifacef.app_nutricionista.repository.AvaliacaoCorporalRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,11 +23,26 @@ public class AvaliacaoCorporalService {
         return repository.findById(id);
     }
 
+    @Transactional
     public AvaliacaoCorporal save(AvaliacaoCorporal avaliacao) {
         return repository.save(avaliacao);
     }
 
-    public void deleteById(Long id) {
-        repository.deleteById(id);
+    @Transactional
+    public AvaliacaoCorporal update(Long id, AvaliacaoCorporal atual) {
+        if (repository.existsById(id)) {
+            atual.setId(id);
+            return repository.save(atual);
+        }
+        return null;
+    }
+
+    @Transactional
+    public boolean deleteById(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
