+++ src/main/java/com/unifacef/app_nutricionista/service/ReceitaService.java
@@ -2,6 +2,7 @@ package com.unifacef.app_nutricionista.service;
 
 import com.unifacef.app_nutricionista.model.Receita;
 import com.unifacef.app_nutricionista.repository.ReceitaRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,11 +23,26 @@ public class ReceitaService {
         return repository.findById(id);
     }
 
+    @Transactional
     public Receita save(Receita receita) {
         return repository.save(receita);
     }
 
-    public void deleteById(Long id) {
-        repository.deleteById(id);
+    @Transactional
+    public Receita update(Long id, Receita atual) {
+        if (repository.existsById(id)) {
+            atual.setId(id);
+            return repository.save(atual);
+        }
+        return null;
+    }
+
+    @Transactional
+    public boolean deleteById(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
